Version 1.4.1
Fixed Some bugs with it showing up as "cidPlayer" hopefully.

Allowed config options to use the old message system.

Optimized the way messages are displayed.

Increased Version Number

Allowed for a softdependency of protocollib. Will use the old message if ProtoocolLib is not on the server.

Removed unnecessary static variables